UPDATE  2023/11/23
ADDED THE TOKEN LOGIN LOG OUT

+++ src/main/java/com/ereceipt/CAZAEORPROJECT/Role/IRoleImplementation.java
@@ -7,10 +7,18 @@ import java.util.List;
 public interface IRoleImplementation {
     List<Role> getAllRoles();
     Role createRole(Role theRole);
-    void deleteRole(Long roleId);
+    void deleteRole(Integer roleId);
+
+
     Role findByName(String name);
-    Role findById(Long roelId);
-    User removeUserFromRole(Long userId, Long roleId);
-    User assignUerToRole(Long userId, Long roleId);
-    Role removeAllUserFromRole(Long roleId);
+
+
+
+    Role findById(Integer roelId);
+
+    User removeUserFromRole(Integer userId, Integer roleId);
+
+    User assignUerToRole(Integer userId, Integer roleId);
+
+    Role removeAllUserFromRole(Integer roleId);
 }
